rsx-auth/pyproject.toml
[project]
name = "rsx-auth"
version = "0.1.0"
description = "OAuth identity service for RSX exchange"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.26",
    "asyncpg>=0.29",
    "pyjwt>=2.8",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rsx_auth"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
